chore: better order for oauth plugin
