+++ apps/core/src/plugins/oauth2/oauth2.ts
@@ -1,12 +1,12 @@
 import { fastifyOauth2 } from '@fastify/oauth2';
-import { UserModel } from '@next/models';
-import { WEB_URL, WEB_URL_LOCAL } from '@next/constants';
-import { Config } from '@/config/config.js';
+import { fastifyPlugin as fp } from 'fastify-plugin';
 import {
   getFacebookUserDetails,
   getGoogleUserDetails,
 } from './utils/get-oauth-info.js';
 import { objectKeys } from './utils/objectKeys.js';
+import { type Querystring, handleOauth } from './handler.js';
+import type { Config } from '@/config/config.js';
 import type { Providers } from '@next/types';
 import type { FastifyInstance, FastifyRequest } from 'fastify';
 
@@ -17,16 +17,7 @@ type NextOauthOptions = {
   facebookSecret: Config['OAUTH_FACEBOOK_SECRET'];
 };
 
-type Query = {
-  // Is User in mobile
-  inApp: string;
-  // The user _id property from mongoose
-  // wouldn't it be better to use the providerId
-  userId: string;
-};
-
-// TODO: implement session token
-export async function oauth2(app: FastifyInstance, opts: NextOauthOptions) {
+async function oauth2(app: FastifyInstance, opts: NextOauthOptions) {
   const providers = {
     google: {
       credentials: {
@@ -54,7 +45,7 @@ export async function oauth2(app: FastifyInstance, opts: NextOauthOptions) {
 
   for (const provider of objectKeys(providers)) {
     const startRedirectPath = `/login/${provider}`;
-    const callbackUri = `http://localhost:5000/v2/login/${provider}/callback`;
+    const callbackUri = `http://localhost:5000/login/${provider}/callback`;
 
     await app.register(fastifyOauth2, {
       name: provider,
@@ -72,63 +63,19 @@ export async function oauth2(app: FastifyInstance, opts: NextOauthOptions) {
 
     app.get(
       `/login/${provider}/callback`,
-      async function (request: FastifyRequest<{ Querystring: Query }>, reply) {
+      async function (
+        request: FastifyRequest<{ Querystring: Querystring }>,
+        reply,
+      ) {
         try {
-          const { inApp = '', userId = '' } = request.query;
-          const { token } =
-            await this[provider].getAccessTokenFromAuthorizationCodeFlow(
-              request,
-            );
-          const oauthUser = await providers[provider].getUserDetails(token);
-          const findUserQuery: Record<string, unknown>[] = [
-            { 'oauth.providerId': oauthUser.providerId },
-          ];
-
-          if (userId) {
-            findUserQuery.push({ _id: userId.split('?')[0] });
-          }
-
-          let user = await UserModel.findOne({
-            $or: findUserQuery,
-          });
-          if (user) {
-            if (userId) {
-              user.set('active', true);
-            }
-            user.set({
-              'oauth.providerId': oauthUser.providerId,
-              'oauth.email': oauthUser.email,
-              'oauth.provider': oauthUser.provider,
-            });
-          } else {
-            user = new UserModel({
-              oauth: {
-                email: oauthUser.email,
-                providerId: oauthUser.providerId,
-                provider: oauthUser.provider,
-              },
-            });
-          }
-
-          await user.save();
-
-          const isLocal =
-            Config.NODE_ENV === 'dev'
-              ? `${WEB_URL_LOCAL}/login?token=${user.generateJWT()}`
-              : `${WEB_URL}/login/token=${user.generateJWT()}`;
-
-          const isMobile =
-            inApp.split('?')[0] === 'true'
-              ? `ufabcnext://login?token=${user.generateJWT()}`
-              : isLocal;
-
-          const redirectTo = isMobile || isLocal;
-          return reply.redirect(redirectTo);
+          await handleOauth.call(this, provider, request, reply, providers);
         } catch (error) {
-          reply.log.error({ error }, 'Error in oauth2');
-          return reply.send(error);
+          reply.log.error({ error: error.data.payload }, 'Error in oauth2');
+          return error.data.payload;
         }
       },
     );
   }
 }
+
+export default fp(oauth2, { name: 'NextOauth2' });
